Move next cloud adapter to its folder
